Project01/pyproject.toml
[project]
name = "2021312779_project01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.4",
    "opencv-python>=4.9.0",
    "pillow>=10.0.0",
    "pycocotools>=2.0.7",
    "pyyaml>=6.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "wandb>=0.25.1",
    "onnx>=1.21.0",
    "onnxscript>=0.7.0",
]
